fixup! CPU: Add op_sw

+++ src/cpu.c
@@ -53,7 +53,7 @@ void op_ori(Cpu *cpu, uint32_t ins) {
   set_reg(cpu, rt, imm | cpu->regs[rs]);
 }
 
-void op_sw(Cpu *cpu, uint8_t ins) {
+void op_sw(Cpu *cpu, uint32_t ins) {
   uint32_t imm_se = get_imm_se(ins);
   uint32_t rs = get_rs(ins);
   uint32_t rt = get_rt(ins);
